[refactor] #53 최적화 연습 첫 번째 N+1 해결

+++ src/main/java/tave/crezipsa/crezipsa/domain/user/entity/UserHistory.java
@@ -9,9 +9,9 @@ import lombok.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class UserHistory {
 
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long historyId;
+	@Id
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private Integer historyId;
 
     @Column(name = "user_id")
     private Long userId;
