+++ src/main/java/tave/crezipsa/crezipsa/domain/user/repository/UserHistoryRepository.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface UserHistoryRepository {
     void save(UserHistory userHistory);
     List<UserHistory> findByUserId(long userId);
-    void deleteByHistoryIdAndUserId(long userId, long historyId);
+    void deleteByHistoryIdAndUserId(long userId, int historyId);
     void deleteByUserId(long userId);
 }
