+++ src/main/java/tave/crezipsa/crezipsa/infrastructure/user/repository/UserHistoryJpaRepository.java
@@ -5,8 +5,8 @@ import tave.crezipsa.crezipsa.domain.user.entity.UserHistory;
 
 import java.util.List;
 
-public interface UserHistoryJpaRepository extends JpaRepository<UserHistory, Long> {
+public interface UserHistoryJpaRepository extends JpaRepository<UserHistory, Integer> {
     List<UserHistory> findTop20ByUserIdOrderByHistoryIdDesc(Long userId);
-    void deleteByUserIdAndHistoryId(Long userId, Long historyId);
+    void deleteByUserIdAndHistoryId(Long userId, Integer historyId);
     void deleteByUserId(Long userId);
 }
